pyproject.toml
[build-system]
requires = ["hatchling>=1.18.0"]
build-backend = "hatchling.build"

[project]
name = "forge-sdlc"
version = "0.1.0"
description = "Forge — Agent-Native SDLC Control Plane powered by BGI Trident"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11"
authors = [
    {name = "Abhinav Khare", email = "abhinav@ahinsa.ai"},
]
keywords = [
    "agentic-ai",
    "sdlc",
    "mcp",
    "orchestration",
    "bgi-trident",
    "developer-tools",
    "ai-governance",
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Software Development :: Build Tools",
    "Topic :: Software Development :: Libraries :: Python Modules",
]

dependencies = [
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
    "anyio>=4.2.0",
    "httpx>=0.26.0",
    "redis>=5.0.0",
    "pyyaml>=6.0.1",
    "tomli>=2.0.1; python_version<'3.11'",
    "networkx>=3.2.1",
    "asyncio-mqtt>=0.16.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "ruff>=0.1.0",
    "mypy>=1.7.0",
    "pre-commit>=3.6.0",
    "mkdocs>=1.5.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.24.0",
]
triclient = [
    "torch>=2.1.0",
    "torch-geometric>=2.4.0",
    "dgl>=2.0.0",
    "xgboost>=2.0.0",
]

[project.urls]
Homepage = "https://github.com/ahinsaai/forge"
Documentation = "https://forge-sdlc.readthedocs.io"
Repository = "https://github.com/ahinsaai/forge"
Issues = "https://github.com/ahinsaai/forge/issues"

[project.scripts]
forge = "forge.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/forge"]

[tool.ruff]
target-version = "py311"
line-length = 100
select = [
    "E",   # pycodestyle errors
    "F",   # Pyflakes
    "I",   # isort
    "N",   # pep8-naming
    "W",   # pycodestyle warnings
    "UP",  # pyupgrade
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "SIM", # flake8-simplify
]
ignore = ["E501"]

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
warn_redundant_casts = true
warn_unused_ignores = true
show_error_codes = true
strict = true

[tool.pytest.ini_options]
minversion = "7.0"
asyncio_mode = "auto"
testpaths = ["tests"]
addopts = "--cov=forge --cov-report=term-missing --cov-report=xml"

[tool.coverage.run]
source = ["src/forge"]
omit = ["*/tests/*", "*/test_*"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]
